requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Vector Database & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# Google Gemini
google-generativeai==0.3.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
